Clean up analyzer behaviour to match actual Reqnroll behaviour
