+++ Reqnroll.Analyzers/AnalyzerReleases.Unshipped.md
@@ -6,7 +6,7 @@
 Rule ID | Category | Severity | Notes
 --------|----------|----------|--------------------
 RR1001  |  Usage   |  Error   | StepTextAnalyzer_StepTextCannotBeNullOrEmptyRule
-RR1002  |  Usage   |  Error   | StepMethodMustReturnVoidOrTaskAnalyzer
-RR1003  |  Usage   |  Error   | AsyncStepMethodMustReturnTaskAnalyzer
+RR1002  |  Usage   | Warning  | StepMethodReturnTypeAnalyzer
+RR1003  |  Usage   |  Error   | AsyncStepMethodReturnTypeAnalyzer
 RR1004  |  Usage   | Warning  | StepTextAnalyzer_StepTextShouldNotHaveLeadingWhitespaceRule
 RR1005  |  Usage   | Warning  | StepTextAnalyzer_StepTextShouldNotHaveTrailingWhitespaceRule
